feat: Organize files
