+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "react-chat",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "description": "React,Firebaseを使用したチャット開発",
   "main": "src/index.js",
   "scripts": {
     "start": "firebase serve",
-    "build": "webpack",
+    "build": "webpack -w",
     "deploy": "firebase deploy",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -22,7 +22,7 @@
   "dependencies": {
     "bufferutil": "4.0.1",
     "firebase": "^5.8.3",
-    "firebase-tools": "^6.3.1",
+    "firebase-tools": "^6.4.0",
     "react": "^16.8.2",
     "react-dom": "^16.8.2",
     "react-redux": "^6.0.1",
